Fix conflict between parameter names and module names in C# code gen.
e.g. In the generated D code for C#, if the wrapped return type is
prefix.Prefix, and the parameter name is prefix, then the compiler
thinks that prefix.Prefix is trying to use UFCS on the parameter instead
of it being the type Prefix in the module prefix. To fix this, the code
now tacks _param onto the end of every parameter name. It's not perfect,
since someone could technically have a module name ending in _param, but
it seems highly unlikely that that will ever actually happen in
practice.