For C#, add _ to the end of member names with the same name as the type.
C# doesn't allow members with the same names as the class or struct that
they're in, because that's how C# names constructors. This would
probably be less of a problem if autowrap didn't try to muck with the
casing of names to match what C# typically uses, but even if that
weren't the case, we can't rely on people using the standard D naming
conventions. As the "simple" tests only have a single field that was a
problem, the portions of this fix which affect other members aren't
tested, but this should at least make the situation better.